fix(Core/Reputations): Faction rep gained by killing mobs is now properly rounded up.
Do not increase reputation rank if exceeds max rank cap.
Fixes #8718

+++ src/server/scripts/Commands/cs_quest.cpp
@@ -307,7 +307,7 @@ public:
                 {
                     if (FactionEntry const* factionEntry = sFactionStore.LookupEntry(repFaction))
                     {
-                        player->GetReputationMgr().SetReputation(factionEntry, repValue);
+                        player->GetReputationMgr().SetReputation(factionEntry, static_cast<float>(repValue));
                     }
                 }
             }
@@ -321,7 +321,7 @@ public:
                 {
                     if (FactionEntry const* factionEntry = sFactionStore.LookupEntry(repFaction))
                     {
-                        player->GetReputationMgr().SetReputation(factionEntry, repValue2);
+                        player->GetReputationMgr().SetReputation(factionEntry, static_cast<float>(repValue2));
                     }
                 }
             }
